fix(apple): Use Task.detached when starting from MainActor (#7766)
When starting a Task, by default it's launched with the same priority as
the calling code.

In the UI these are run on the `MainActor` with highest priority by
default. If the worker thread running the Task closure gets blocked, it
will cause the UI to hang.

To fix this, we use `Task.detached` which runs the closure without a
specific priority, which is lower than the UI thread.

Furthermore, `weak self` is used to prevent retain cycles if the parent
thread `deinit`s.

This was causing an issue primarily when making IPC calls because those
will sometimes hang until the XPC service is launched for the first
time.

---------

+++ swift/apple/FirezoneKit/Sources/FirezoneKit/Views/MenuBar.swift
@@ -265,13 +265,15 @@ public final class MenuBar: NSObject, ObservableObject {
   }
 
   @objc private func signOutButtonTapped() {
-    Task {
-      try await model.store.signOut()
+    Task.detached { [weak self] in
+      try await self?.model.store.signOut()
     }
   }
 
   @objc private func grantPermissionMenuItemTapped() {
-    Task {
+    Task.detached { [weak self] in
+      guard let self else { return }
+
       do {
         // If we get here, it means either system extension got disabled or
         // our VPN configuration got removed. Since we don't know which, reinstall
@@ -314,9 +316,11 @@ public final class MenuBar: NSObject, ObservableObject {
   }
 
   @objc private func quitButtonTapped() {
-    Task {
-      model.store.stop()
-      NSApp.terminate(self)
+    Task.detached { [weak self] in
+      guard let self else { return }
+
+      await self.model.store.stop()
+      await NSApp.terminate(self)
     }
   }
 
@@ -357,8 +361,8 @@ public final class MenuBar: NSObject, ObservableObject {
     guard connectingAnimationTimer == nil else { return }
     let timer = Timer(timeInterval: 0.25, repeats: true) { [weak self] _ in
       guard let self = self else { return }
-      Task {
-        await self.connectingAnimationShowNextFrame()
+      Task.detached { [weak self] in
+        await self?.connectingAnimationShowNextFrame()
       }
     }
     RunLoop.main.add(timer, forMode: .common)
